third_party/blink/web_tests/fragmentation/auto-scrollbar-shrink-to-fit.html

<!DOCTYPE html>
<style>
    /* Prevent double initial layout. */
    body { overflow:scroll; }
</style>
<p>There should be a blue square below.</p>
<div style="columns:2; column-fill:auto; gap:0; width:200px; height:110px; line-height:20px; orphans:1; widows:1;">
    <div style="position:relative;">
        <div style="height:80px;"></div>
        <div style="position:absolute;">
            <div><br></div>
            <!-- #elm doesn't fit in the first column (only 10px left, and it needs 20px),
                 so it has to be pushed to the next one. -->
            <div id="elm" style="width:20px; background:blue;"><br></div>
            <div style="visibility:hidden; overflow-y:auto; height:30px;">
                <br>
                <br>
            </div>
        </div>
    </div>
</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(() => {
    var elm = document.getElementById("elm");
    assert_equals(elm.offsetTop, -80);
    assert_equals(elm.offsetLeft, 100);
}, "Shrink-to-fit with auto vertical scrollbar inside");
</script>
